End of Section 5

+++ src/AuctionService/Controllers/AuctionController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -47,12 +48,13 @@ public class AuctionController : ControllerBase
 
         return _mapper.Map<AuctionDTO>(auction);
     }
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDTO>> CreateAuction(CreateAuctionDTO auctionDTO)
     {
         var auction = _mapper.Map<Auction>(auctionDTO);
         // TODO: Add current user as seller
-        auction.Seller = "test";
+        auction.Seller = User.Identity.Name;
 
         _context.Auctions.Add(auction);
         
@@ -67,6 +69,7 @@ public class AuctionController : ControllerBase
         return CreatedAtAction(nameof(GetAuctionById), 
             new {auction.Id}, _mapper.Map<AuctionDTO>(auction));
     }
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDTO updateAuctionDTO)
     {
@@ -75,7 +78,7 @@ public class AuctionController : ControllerBase
         
         if (auction == null) return NotFound();
 
-        // TODO: Add current user as seller
+        if(auction.Seller != User.Identity.Name) return Forbid();
 
         auction.Item.Make = updateAuctionDTO.Make ?? auction.Item.Make;    
         auction.Item.Model = updateAuctionDTO.Model ?? auction.Item.Model;    
@@ -90,7 +93,7 @@ public class AuctionController : ControllerBase
         if (result) return Ok();
         return BadRequest("Could not save changes to the DB");
     }
-
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -98,7 +101,7 @@ public class AuctionController : ControllerBase
 
         if(auction == null) return NotFound();
 
-        // TODO: Add current user as seller
+        if(auction.Seller != User.Identity.Name) return Forbid();
 
         _context.Auctions.Remove(auction);
